requirements.txt
discord.py==1.7.2
python-dotenv==0.17.1
discord-py-slash-command==1.1.2